migrate dzz from SS.ABT

+++ src/AlgoTradeForge.Domain/Trading/Order.cs
@@ -7,9 +7,9 @@ public sealed class Order
     public required OrderSide Side { get; init; }
     public required OrderType Type { get; init; }
     public required decimal Quantity { get; init; }
-    public decimal? LimitPrice { get; init; }
-    public decimal? StopPrice { get; init; }
-    public decimal? StopLossPrice { get; init; }
+    public long? LimitPrice { get; init; }
+    public long? StopPrice { get; init; }
+    public long? StopLossPrice { get; init; }
     public IReadOnlyList<TakeProfitLevel>? TakeProfitLevels { get; init; }
     public OrderStatus Status { get; internal set; } = OrderStatus.Pending;
     public DateTimeOffset SubmittedAt { get; internal set; }
