+++ tests/AlgoTradeForge.Domain.Tests/Engine/BarMatcherTests.cs
@@ -50,7 +50,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_LimitBuyPriceAtOrAboveLow_Fills()
     {
-        var order = TestOrders.LimitBuy(TestAssets.Aapl, 100m, limitPrice: 14800m);
+        var order = TestOrders.LimitBuy(TestAssets.Aapl, 100m, limitPrice: 14800L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200);
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -63,7 +63,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_LimitBuyPriceBelowLow_Rejects()
     {
-        var order = TestOrders.LimitBuy(TestAssets.Aapl, 100m, limitPrice: 14700m);
+        var order = TestOrders.LimitBuy(TestAssets.Aapl, 100m, limitPrice: 14700L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200);
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -75,7 +75,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_LimitSellPriceAtOrBelowHigh_Fills()
     {
-        var order = TestOrders.LimitSell(TestAssets.Aapl, 100m, limitPrice: 15500m);
+        var order = TestOrders.LimitSell(TestAssets.Aapl, 100m, limitPrice: 15500L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200);
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -88,7 +88,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_LimitSellPriceAboveHigh_Rejects()
     {
-        var order = TestOrders.LimitSell(TestAssets.Aapl, 100m, limitPrice: 15600m);
+        var order = TestOrders.LimitSell(TestAssets.Aapl, 100m, limitPrice: 15600L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200);
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -166,7 +166,7 @@ public class BarMatcherTests
     public void TryFill_StopBuy_TriggersWhenHighReachesStopPrice()
     {
         // Stop Buy triggers when price rises to stop price (breakout entry)
-        var order = TestOrders.StopBuy(TestAssets.Aapl, 100m, stopPrice: 15200m);
+        var order = TestOrders.StopBuy(TestAssets.Aapl, 100m, stopPrice: 15200L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200); // High=15500 >= Stop=15200
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -179,7 +179,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_StopBuy_DoesNotTriggerWhenHighBelowStop()
     {
-        var order = TestOrders.StopBuy(TestAssets.Aapl, 100m, stopPrice: 16000m);
+        var order = TestOrders.StopBuy(TestAssets.Aapl, 100m, stopPrice: 16000L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200); // High=15500 < Stop=16000
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -192,7 +192,7 @@ public class BarMatcherTests
     public void TryFill_StopBuy_GapsUp_FillsAtOpen()
     {
         // When bar opens above stop price, fill at open (gap up)
-        var order = TestOrders.StopBuy(TestAssets.Aapl, 100m, stopPrice: 15000m);
+        var order = TestOrders.StopBuy(TestAssets.Aapl, 100m, stopPrice: 15000L);
         var bar = TestBars.Create(15500, 16000, 15400, 15800); // Open=15500 > Stop=15000
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -206,7 +206,7 @@ public class BarMatcherTests
     public void TryFill_StopSell_TriggersWhenLowReachesStopPrice()
     {
         // Stop Sell triggers when price falls to stop price (breakdown entry)
-        var order = TestOrders.StopSell(TestAssets.Aapl, 100m, stopPrice: 14900m);
+        var order = TestOrders.StopSell(TestAssets.Aapl, 100m, stopPrice: 14900L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200); // Low=14800 <= Stop=14900
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -219,7 +219,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_StopSell_DoesNotTriggerWhenLowAboveStop()
     {
-        var order = TestOrders.StopSell(TestAssets.Aapl, 100m, stopPrice: 14000m);
+        var order = TestOrders.StopSell(TestAssets.Aapl, 100m, stopPrice: 14000L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200); // Low=14800 > Stop=14000
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -231,7 +231,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_StopSell_GapsDown_FillsAtOpen()
     {
-        var order = TestOrders.StopSell(TestAssets.Aapl, 100m, stopPrice: 15500m);
+        var order = TestOrders.StopSell(TestAssets.Aapl, 100m, stopPrice: 15500L);
         var bar = TestBars.Create(15000, 15200, 14800, 14900); // Open=15000 < Stop=15500
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -244,7 +244,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_StopBuy_WithSlippage_FillsAtStopPlusSlippage()
     {
-        var order = TestOrders.StopBuy(TestAssets.Aapl, 100m, stopPrice: 15200m);
+        var order = TestOrders.StopBuy(TestAssets.Aapl, 100m, stopPrice: 15200L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200);
         var options = CreateOptions(TestAssets.Aapl, slippageTicks: 5m); // 5 * 0.01 = 0.05
 
@@ -257,7 +257,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_StopSell_WithSlippage_FillsAtStopMinusSlippage()
     {
-        var order = TestOrders.StopSell(TestAssets.Aapl, 100m, stopPrice: 14900m);
+        var order = TestOrders.StopSell(TestAssets.Aapl, 100m, stopPrice: 14900L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200);
         var options = CreateOptions(TestAssets.Aapl, slippageTicks: 5m);
 
@@ -275,21 +275,21 @@ public class BarMatcherTests
     public void TryFill_StopLimitBuy_TriggersAndFillsWhenLimitInRange()
     {
         // StopLimit Buy: stop triggers at 15200, then limit buy at 15200 if price allows
-        var order = TestOrders.StopLimitBuy(TestAssets.Aapl, 100m, stopPrice: 15200m, limitPrice: 15300m);
+        var order = TestOrders.StopLimitBuy(TestAssets.Aapl, 100m, stopPrice: 15200L, limitPrice: 15300L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200); // High >= Stop, Limit <= High
         var options = CreateOptions(TestAssets.Aapl);
 
         var fill = _matcher.TryFill(order, bar, options);
 
         Assert.NotNull(fill);
-        Assert.Equal(15300m, fill.Price); // Fill at limit price
+        Assert.Equal(15300L, fill.Price); // Fill at limit price
     }
 
     [Fact]
     public void TryFill_StopLimitBuy_TriggersButLimitNotReached_Pending()
     {
         // Stop triggers but limit price not reached — order becomes Triggered (pending limit)
-        var order = TestOrders.StopLimitBuy(TestAssets.Aapl, 100m, stopPrice: 15200m, limitPrice: 14700m);
+        var order = TestOrders.StopLimitBuy(TestAssets.Aapl, 100m, stopPrice: 15200L, limitPrice: 14700L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200); // High >= Stop, but Low > Limit
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -303,7 +303,7 @@ public class BarMatcherTests
     public void TryFill_StopLimitBuy_AlreadyTriggered_FillsAsLimit()
     {
         // Once triggered, StopLimit acts as a regular Limit order
-        var order = TestOrders.StopLimitBuy(TestAssets.Aapl, 100m, stopPrice: 15200m, limitPrice: 14900m);
+        var order = TestOrders.StopLimitBuy(TestAssets.Aapl, 100m, stopPrice: 15200L, limitPrice: 14900L);
         order.Status = OrderStatus.Triggered;
         var bar = TestBars.Create(15000, 15500, 14800, 15200); // Low=14800 <= Limit=14900
         var options = CreateOptions(TestAssets.Aapl);
@@ -317,7 +317,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_StopLimitSell_TriggersAndFillsWhenLimitInRange()
     {
-        var order = TestOrders.StopLimitSell(TestAssets.Aapl, 100m, stopPrice: 14900m, limitPrice: 14800m);
+        var order = TestOrders.StopLimitSell(TestAssets.Aapl, 100m, stopPrice: 14900L, limitPrice: 14800L);
         var bar = TestBars.Create(15000, 15500, 14700, 15200); // Low <= Stop, Limit >= Low
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -330,7 +330,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_StopLimitSell_TriggersButLimitNotReached_Pending()
     {
-        var order = TestOrders.StopLimitSell(TestAssets.Aapl, 100m, stopPrice: 14900m, limitPrice: 15600m);
+        var order = TestOrders.StopLimitSell(TestAssets.Aapl, 100m, stopPrice: 14900L, limitPrice: 15600L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200); // Low <= Stop, but High < Limit
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -343,7 +343,7 @@ public class BarMatcherTests
     [Fact]
     public void TryFill_StopLimitBuy_StopNotReached_NothingHappens()
     {
-        var order = TestOrders.StopLimitBuy(TestAssets.Aapl, 100m, stopPrice: 16000m, limitPrice: 16100m);
+        var order = TestOrders.StopLimitBuy(TestAssets.Aapl, 100m, stopPrice: 16000L, limitPrice: 16100L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200); // High < Stop
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -357,7 +357,7 @@ public class BarMatcherTests
     public void TryFill_StopLimitBuy_NoSlippageOnLimitFill()
     {
         // StopLimit orders fill at limit price — no slippage
-        var order = TestOrders.StopLimitBuy(TestAssets.Aapl, 100m, stopPrice: 15200m, limitPrice: 15300m);
+        var order = TestOrders.StopLimitBuy(TestAssets.Aapl, 100m, stopPrice: 15200L, limitPrice: 15300L);
         var bar = TestBars.Create(15000, 15500, 14800, 15200);
         var options = CreateOptions(TestAssets.Aapl, slippageTicks: 10m);
 
@@ -371,7 +371,7 @@ public class BarMatcherTests
 
     #region SL/TP Evaluation Tests (T030)
 
-    private static Order BuyOrderWithSlTp(decimal slPrice, params TakeProfitLevel[] tpLevels) =>
+    private static Order BuyOrderWithSlTp(long slPrice, params TakeProfitLevel[] tpLevels) =>
         new()
         {
             Id = 1,
@@ -383,7 +383,7 @@ public class BarMatcherTests
             TakeProfitLevels = tpLevels.Length > 0 ? tpLevels : null,
         };
 
-    private static Order SellOrderWithSlTp(decimal slPrice, params TakeProfitLevel[] tpLevels) =>
+    private static Order SellOrderWithSlTp(long slPrice, params TakeProfitLevel[] tpLevels) =>
         new()
         {
             Id = 2,
@@ -398,7 +398,7 @@ public class BarMatcherTests
     [Fact]
     public void EvaluateSlTp_BuySide_SlOnlyHit_ReturnsSlFill()
     {
-        var order = BuyOrderWithSlTp(slPrice: 14500m, new TakeProfitLevel(16000m, 1m));
+        var order = BuyOrderWithSlTp(slPrice: 14500L, new TakeProfitLevel(16000L, 1m));
         // Bar drops to SL but doesn't reach TP
         var bar = TestBars.Create(15000, 15200, 14400, 14800); // Low=14400 <= SL=14500, High=15200 < TP=16000
         var options = CreateOptions(TestAssets.Aapl);
@@ -414,7 +414,7 @@ public class BarMatcherTests
     [Fact]
     public void EvaluateSlTp_BuySide_TpOnlyHit_ReturnsTpFill()
     {
-        var order = BuyOrderWithSlTp(slPrice: 14500m, new TakeProfitLevel(15800m, 1m));
+        var order = BuyOrderWithSlTp(slPrice: 14500L, new TakeProfitLevel(15800L, 1m));
         // Bar rises to TP but doesn't drop to SL
         var bar = TestBars.Create(15000, 16000, 14600, 15500); // High=16000 >= TP=15800, Low=14600 > SL=14500
         var options = CreateOptions(TestAssets.Aapl);
@@ -431,7 +431,7 @@ public class BarMatcherTests
     [Fact]
     public void EvaluateSlTp_BuySide_BothInRange_WorstCase_ReturnsSl()
     {
-        var order = BuyOrderWithSlTp(slPrice: 14500m, new TakeProfitLevel(15800m, 1m));
+        var order = BuyOrderWithSlTp(slPrice: 14500L, new TakeProfitLevel(15800L, 1m));
         // Bar covers both SL and TP
         var bar = TestBars.Create(15000, 16000, 14000, 15500); // Low=14000 <= SL, High=16000 >= TP
         var options = CreateOptions(TestAssets.Aapl);
@@ -445,7 +445,7 @@ public class BarMatcherTests
     [Fact]
     public void EvaluateSlTp_BuySide_NeitherHit_ReturnsNull()
     {
-        var order = BuyOrderWithSlTp(slPrice: 14000m, new TakeProfitLevel(16000m, 1m));
+        var order = BuyOrderWithSlTp(slPrice: 14000L, new TakeProfitLevel(16000L, 1m));
         var bar = TestBars.Create(15000, 15200, 14100, 15100); // Low=14100 > SL=14000, High=15200 < TP=16000
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -457,7 +457,7 @@ public class BarMatcherTests
     [Fact]
     public void EvaluateSlTp_SellSide_SlOnlyHit_ReturnsSlFill()
     {
-        var order = SellOrderWithSlTp(slPrice: 15500m, new TakeProfitLevel(14000m, 1m));
+        var order = SellOrderWithSlTp(slPrice: 15500L, new TakeProfitLevel(14000L, 1m));
         // Short position: SL is above entry, bar rises to SL
         var bar = TestBars.Create(15000, 15600, 14200, 15300); // High=15600 >= SL=15500, Low=14200 > TP=14000
         var options = CreateOptions(TestAssets.Aapl);
@@ -472,7 +472,7 @@ public class BarMatcherTests
     [Fact]
     public void EvaluateSlTp_SellSide_TpOnlyHit_ReturnsTpFill()
     {
-        var order = SellOrderWithSlTp(slPrice: 15500m, new TakeProfitLevel(14200m, 1m));
+        var order = SellOrderWithSlTp(slPrice: 15500L, new TakeProfitLevel(14200L, 1m));
         // Short position: TP is below entry, bar drops to TP
         var bar = TestBars.Create(15000, 15400, 14100, 14500); // Low=14100 <= TP=14200, High=15400 < SL=15500
         var options = CreateOptions(TestAssets.Aapl);
@@ -486,7 +486,7 @@ public class BarMatcherTests
     [Fact]
     public void EvaluateSlTp_ZeroRangeBar_NoSlTpHit()
     {
-        var order = BuyOrderWithSlTp(slPrice: 14000m, new TakeProfitLevel(16000m, 1m));
+        var order = BuyOrderWithSlTp(slPrice: 14000L, new TakeProfitLevel(16000L, 1m));
         var bar = TestBars.Create(15000, 15000, 15000, 15000); // Zero-range bar
         var options = CreateOptions(TestAssets.Aapl);
 
@@ -499,9 +499,9 @@ public class BarMatcherTests
     public void EvaluateSlTp_MultipleTpLevels_PartialClosure()
     {
         var order = BuyOrderWithSlTp(
-            slPrice: 14000m,
-            new TakeProfitLevel(15500m, 0.5m),  // TP1: 50% at 15500
-            new TakeProfitLevel(16000m, 1m));    // TP2: remaining at 16000
+            slPrice: 14000L,
+            new TakeProfitLevel(15500L, 0.5m),  // TP1: 50% at 15500
+            new TakeProfitLevel(16000L, 1m));    // TP2: remaining at 16000
 
         // First TP hit
         var bar1 = TestBars.Create(15000, 15600, 14100, 15300); // High=15600 >= TP1=15500
